% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/content.R
\name{find_parenthesis}
\alias{find_parenthesis}
\title{Helper to find first block of parenthesis}
\usage{
find_parenthesis(text)
}
\arguments{
\item{text}{the string/code/statement you want to parse.}
}
\value{
a substring. Either "" or the first parenthesis block.
}
\description{
This function will search for the first block of parenthesis and return it if found. Otherwise, it will return "".
}
\examples{
# returns ""
find_parenthesis("3 + 5")
# returns "(3 + 5)"
find_parenthesis("2 * (3 + 5)")
}
