% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/content.R
\name{map_ref}
\alias{map_ref}
\title{arguments to use with @inheritParams}
\arguments{
\item{map}{a named list of data.frames/lists/matrices. Where names are keys for referencing the values in the \code{text} parameters.}
}
\description{
arguments to use with @inheritParams
}
\keyword{internal}
