% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/content.R
\name{mapping_names_ref}
\alias{mapping_names_ref}
\title{arguments to use with @inheritParams}
\arguments{
\item{mapping_names}{optional argument to make the function faster or limit which map elements can be referenced.}
}
\description{
arguments to use with @inheritParams
}
\keyword{internal}
