% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/content.R
\name{text_ref}
\alias{text_ref}
\title{arguments to use with @inheritParams}
\arguments{
\item{text}{the string/code/statement you want to parse.}
}
\description{
arguments to use with @inheritParams
}
\keyword{internal}
