% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/content.R
\name{tree_ref}
\alias{tree_ref}
\title{arguments to use with @inheritParams}
\arguments{
\item{tree}{the output object from \link{create_tree}}
}
\description{
arguments to use with @inheritParams
}
\keyword{internal}
